Code refactoring, moving some common functions to external class. Changed the database approach for user avatars loading via /api/user{id}/avatar endpoint (new format is {filename} instead of {folder/filename} as FileController kinda forces that with its method to read the files from folder [defined in application.properties]). Started changing types of functions to ResponseEntity<> for more clarity for front-end applications (with more precise messages [i.e. what went wrong])

+++ src/main/java/inz/gameadvisor/restapi/service/DevicesService.java
@@ -1,34 +1,28 @@
 package inz.gameadvisor.restapi.service;
 
+import inz.gameadvisor.restapi.misc.CustomFunctions;
 import inz.gameadvisor.restapi.misc.CustomRepsonses;
 import inz.gameadvisor.restapi.model.deviceOriented.*;
 import inz.gameadvisor.restapi.model.userOriented.User;
-import inz.gameadvisor.restapi.model.userOriented.UserShow;
 import inz.gameadvisor.restapi.repository.*;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.apache.tomcat.util.codec.binary.Base64;
-import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityManager;
-import javax.persistence.NoResultException;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
+import javax.persistence.*;
 import javax.transaction.Transactional;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class DevicesService {
+public class DevicesService extends CustomFunctions {
 
     private final DevicesRepository devicesRepository;
     private final CPURepository cpuRepository;
@@ -36,6 +30,10 @@ public class DevicesService {
     private final RAMRepository ramRepository;
     private final OSRepository osRepository;
     private final UserRepository userRepository;
+
+    //private final CustomFunctions customFunctions = new CustomFunctions();
+    String path;
+    String message;
 
     @PersistenceContext
     EntityManager em;
@@ -58,7 +56,7 @@ public class DevicesService {
         }
     }
 
-    public void addDevice(UpdatedDevices device, String token){
+    public ResponseEntity<Object> addDevice(UpdatedDevices device, String token){
 
         long userID = getUserIDFromToken(token);
 
@@ -83,26 +81,38 @@ public class DevicesService {
             devicesRepository.save(createdDevice);
         }
         catch (DataIntegrityViolationException e){
-            throw new CustomRepsonses.MyDataConflict("Couldn't create device");
+            return responseFromServer(HttpStatus.CONFLICT,path,message);
         }
+        return null;
     }
 
     @SneakyThrows
-    public void deleteDevice(long id, String token){
+    public ResponseEntity<Object> deleteDevice(long id, String token){
         long userID = getUserIDFromToken(token);
+        path = "/api/device/" + id+ "/delete";
 
-        Devices device = devicesRepository.findById(id).orElseThrow(() -> new CustomRepsonses.MyNotFoundException("Device of ID: " + id + " not found."));
+        Optional<Devices> device = devicesRepository.findById(id);
 
-        User user = device.getUser();
+        if(device.isEmpty()){
+            message = "No such device was found";
+            return responseFromServer(HttpStatus.NOT_FOUND,path,message);
+        }
+
+        User user = device.get().getUser();
 
         if(userID == user.getUserID()){
             devicesRepository.deleteById(id);
+            message = "Device deleted";
+            return responseFromServer(HttpStatus.OK,path,message);
         }
         else if(isUserAnAdmin(userID)){
             devicesRepository.deleteById(id);
+            message = "Device deleted";
+            return responseFromServer(HttpStatus.OK,path,message);
         }
         else{
-            throw new CustomRepsonses.MyForbiddenAccess("User of ID " + userID + " tried to delete device of (ID): " + user.getUserID());
+            message = "You don't have access to that";
+            return responseFromServer(HttpStatus.FORBIDDEN,path,message);
         }
     }
 
@@ -110,93 +120,71 @@ public class DevicesService {
     @Transactional
     public ResponseEntity<Object> editDevice(UpdatedDevices updatedDevices, long id, String token){
         long userID = getUserIDFromToken(token);
+        path = "/api/device/" + id + "/edit";
 
-        Devices device = devicesRepository.findById(id).orElseThrow(() -> new CustomRepsonses.MyNotFoundException("Device of ID: " + id + " not found."));
+        if(Objects.isNull(updatedDevices.getShortName()) || updatedDevices.getShortName().isBlank()){
+            message = "Bad request";
+            return responseFromServer(HttpStatus.BAD_REQUEST,path,message);
+        }
 
-        User user = device.getUser();
+        Optional<Devices> device = devicesRepository.findById(id);
+        if(device.isEmpty()){
+            message = "No such device found";
+            return responseFromServer(HttpStatus.NOT_FOUND,path,message);
+        }
 
-        LinkedHashMap<String, String> jsonOrderedMap = new LinkedHashMap<String, String>();
-        JSONObject response = new JSONObject(jsonOrderedMap);
+        User user = device.get().getUser();
+
+        String shortName = updatedDevices.getShortName();
+        long cpuID = updatedDevices.getCpuID();
+        Optional<CPU> cpu = cpuRepository.findById(cpuID);
+        if(cpu.isEmpty()){
+            message = "No such CPU was found";
+            return responseFromServer(HttpStatus.NOT_FOUND,path,message);
+        }
+
+        long gpuID = updatedDevices.getGpuID();
+        long ramID = updatedDevices.getRamID();
+        long osID = updatedDevices.getOsID();
+        boolean isHDD = updatedDevices.isHDD();
+        boolean isSSD = updatedDevices.isSSD();
 
         if(userID == user.getUserID()){
-            String shortName = updatedDevices.getShortName();
-            long cpuID = updatedDevices.getCpuID();
-            long gpuID = updatedDevices.getGpuID();
-            long ramID = updatedDevices.getRamID();
-            long osID = updatedDevices.getOsID();
-            boolean isHDD = updatedDevices.isHDD();
-            boolean isSSD = updatedDevices.isSSD();
-
             if(!shortName.isBlank()){
-                if(!checkIfRecordWithSameNameExists("shortName","devices", shortName))
+                if(!checkIfSameRecordExists("shortName","devices", shortName))
                 {
-                    Query query = em.createNativeQuery("UPDATE devices SET shortName = ? WHERE deviceID = ?")
-                            .setParameter(1, shortName)
-                            .setParameter(2, id);
-                    query.executeUpdate();
-                    System.out.println("Update executed");
+                    if(updateField("devices","shortName",shortName,"deviceID",String.valueOf(id)) == 0){
+                        message = "Updating was not successful";
+                        return responseFromServer(HttpStatus.INTERNAL_SERVER_ERROR,path,message);
+                    }
                 }
                 else{
-                    Date date = new Date(System.currentTimeMillis());
-                    response.put("message","Device with such name exists");
-                    response.put("code", HttpStatus.CONFLICT.value());
-                    response.put("timestamp",date);
-                    response.put("path","/api/device/"+id+"/edit");
-                    return new ResponseEntity<>(response.toMap(), HttpStatus.CONFLICT);
+                    message = "One of your devices with such name exists";
+                    return responseFromServer(HttpStatus.CONFLICT,path,message);
                 }
+            }
+            if(updateField("devices","cpuID",String.valueOf(cpuID),"deviceID",String.valueOf(id)) == 0){
+                message = "Updating was not successful";
+                return responseFromServer(HttpStatus.INTERNAL_SERVER_ERROR,path,message);
+            }
+            message = "Device has been modified";
+            return responseFromServer(HttpStatus.OK,path,message);
+        }
+        else if(isUserAnAdmin(userID)){
+            if(!shortName.isBlank()){
+                Query query = em.createNativeQuery("UPDATE devices SET shortName = ? WHERE deviceID = ?")
+                        .setParameter(1, shortName)
+                        .setParameter(2, id);
+                query.executeUpdate();
+                message = "Device has been modified";
+                return responseFromServer(HttpStatus.OK,path,message);
             }
         }
         else{
-            return new ResponseEntity<>("You don't have access to that resource", new HttpHeaders(), HttpStatus.FORBIDDEN);
+            message = "You don't have access to that resource";
+            return responseFromServer(HttpStatus.FORBIDDEN,path,message);
         }
-        return new ResponseEntity<>("Nothing changed", new HttpHeaders(), HttpStatus.NOT_MODIFIED);
-    }
-
-    //Token value for testing
-    //eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1c2VyMkB1c2VyLmNvbSIsInJvbGVzIjoiUk9MRV9VU0VSIiwiZXhwIjoxNjM5NTk3OTg1LCJ1c2VySUQiOjJ9.V07n9xBSrGCLYn849ryVsZnAkmHQDjEAxNjKxpafjOw
-
-
-    public JSONObject getBodyFromToken(String token){
-        String[] splitString = token.split("\\.");
-        String base64EncodedBody = splitString[1];
-        Base64 base64Url = new Base64(true);
-        String body = new String(base64Url.decode(base64EncodedBody));
-
-        return new JSONObject(body);
-    }
-
-    public boolean checkIfRecordWithSameNameExists(String columnName,String tableName, String name){
-
-        Query query = em.createNativeQuery("SELECT " + columnName + " FROM " + tableName + " WHERE " + columnName+ " = ?")
-                .setParameter(1, name);
-        try{
-            query.getSingleResult();
-        }
-        catch (NoResultException e){
-            e.getLocalizedMessage();
-            return false;
-        }
-        return true;
-    }
-
-    public long getUserIDFromToken(String token){
-        JSONObject tokenBody = getBodyFromToken(token);
-
-        return Long.parseLong(tokenBody.get("userID").toString());
-    }
-
-    public String getUserRoleFromToken(String token){
-        JSONObject tokenBody = getBodyFromToken(token);
-
-        return tokenBody.get("roles").toString();
-    }
-
-    public boolean isUserAnAdmin(long userID) {
-        Query query = em.createNativeQuery("SELECT roles FROM users WHERE userID = ?;")
-                .setParameter(1, userID);
-
-        String queryUserRole = query.getSingleResult().toString();
-
-        return queryUserRole.equals("ROLE_ADMIN");
+        message = "Method has not been applied";
+        return responseFromServer(HttpStatus.NOT_ACCEPTABLE,path,message);
     }
 }
